Tries to make g++ nicer to randomseed()'s reinterpret_cast

+++ nativefuncs/math.cpp
@@ -389,7 +389,7 @@ void Program::construct_math_library()
 		if(!args.size())
 			return Value(Value::vType::Null, int(ErrorCode::NotEnoughArgs));
 
-		srand(reinterpret_cast<unsigned int>(args[0].t_value.as_object_ptr)); // Who cares
+		srand(reinterpret_cast<uint64_t>(args[0].t_value.as_object_ptr)); // Who cares
 
 		return Value();
 	}));
